Clear command, backup media command, minor exception rethrowing addition, removal of BackupChannelConductor as redundant

+++ src/main/java/com/multiheaded/vladikbot/utils/FileUtils.java
@@ -1,26 +1,32 @@
 package com.multiheaded.vladikbot.utils;
 
+import java.io.BufferedWriter;
 import java.io.File;
+import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
-import java.util.Objects;
+import java.nio.file.Paths;
+import java.util.Set;
 
 /**
  * @author Oliver Johnson
  */
 public class FileUtils {
-    public static File getFileByIdAndExtension(String pathToDir, String id, String extension) {
+    public static File getFileByIdAndExtension(String pathToDir, String id, String extension) throws IOException {
+        Files.createDirectories(Paths.get(pathToDir));
         File folder = new File(pathToDir);
         File[] paths = folder.listFiles();
         File file = null;
 
-        for (File path : Objects.requireNonNull(paths)) {
-            if (path.toString().contains(id) && path.toString().contains(extension)) {
-                if (file != null) {
-                    file = (file.lastModified() > path.lastModified()) ? file : path;
-                } else {
-                    file = path;
+        if (paths != null) {
+            for (File path : paths) {
+                if (path.toString().contains(id) && path.toString().contains(extension)) {
+                    if (file != null) {
+                        file = (file.lastModified() > path.lastModified()) ? file : path;
+                    } else {
+                        file = path;
+                    }
                 }
             }
         }
@@ -37,11 +43,24 @@ public class FileUtils {
         File f = new File(pathToDir);
         File[] paths = f.listFiles();
 
-        for(File path : Objects.requireNonNull(paths)) {
-            if (path.toString().contains(id) && path.toString().contains(extension))
-                if (!path.delete()) {
-                    throw new SecurityException();
-                }
+        if (paths != null) {
+            for (File path : paths) {
+                if (path.toString().contains(id) && path.toString().contains(extension))
+                    if (!path.delete()) {
+                        throw new SecurityException();
+                    }
+            }
         }
+    }
+
+    public static void writeSetToFile(String pathToFile, Set<String> set) throws IOException {
+        BufferedWriter out = new BufferedWriter(new FileWriter(pathToFile));
+
+        for (String raw : set) {
+            out.write(raw);
+            out.newLine();
+        }
+
+        out.close();
     }
 }
